#68 feat(fe): upload버튼 부분 트리거 코드 분리
